refactor: clean up settings menu with simpler names, reorganized global namespaced code
