+++ assets/translations/en.json
@@ -11,10 +11,10 @@
   "title-view-create-group-public": "Create A Public Group",
   "title-view-devices": "Devices",
   "title-view-settings": "Settings",
-  "title-view-theming": "Theming",
+  "title-view-theming": "Appearance",
   "title-view-advanced": "Advanced",
   "title-view-invite": "Invite",
-  "title-view-preferences-chat": "Chat Preferences",
+  "title-view-preferences-chat": "Chats",
   "title-view-privacy": "Privacy",
   "title-view-homeserver-search": "Find Your Homeserve",
   "title-view-profile": "Set up Your Profile",
@@ -53,8 +53,8 @@
   "button-text-load-captcha": "Load Captcha",
   "button-text-confirmed": "Confirmed",
   "list-item-settings-notification": "Notifications",
-  "list-item-settings-chat": "Chats and Media",
-  "list-item-settings-privacy": "Security & Privacy",
+  "list-item-settings-chat": "Chats",
+  "list-item-settings-privacy": "Privacy",
   "list-item-settings-sms": "SMS and MMS",
   "list-item-settings-logout": "Logout",
   "alert-restart-app-effect": "You'll need to close and restart the app for this to take affect",
