+++ lib/views/home/settings/devices.dart
@@ -1,11 +1,7 @@
-// Dart imports:
-import 'dart:async';
-
 // Flutter imports:
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 
 // Package imports:
 import 'package:equatable/equatable.dart';
@@ -16,9 +12,7 @@ import 'package:redux/redux.dart';
 // Project imports:
 import 'package:syphon/global/colours.dart';
 import 'package:syphon/global/dimensions.dart';
-import 'package:syphon/global/strings.dart';
 import 'package:syphon/store/auth/actions.dart';
-import 'package:syphon/store/crypto/actions.dart';
 import 'package:syphon/store/index.dart';
 import 'package:syphon/store/settings/actions.dart';
 import 'package:syphon/store/settings/devices-settings/model.dart';
